+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [v.15.1]
-- stable 
+- stable
+- Added `utils.extract_keys_with_all_words`
 
 ## [v0.15.0]
 - `platforms_schemas` update